Fixed failing tests. Removed euclid as a dependency

+++ examples/simple.rs
@@ -12,8 +12,8 @@ fn main() {
     let mut viewport = Viewport::new(ScreenPos::new(0, 4), viewport_size);
 
     // Camera
-    let (width, height) = (width as f32, height as f32);
-    let camera_size = WorldSize::new(width / 2.0, height / 2.0);
+    let (width, height) = (width as i64, height as i64);
+    let camera_size = WorldSize::new(width / 2, height / 2);
     let camera_pos = WorldPos::new(width, height);
     let camera = Camera::new(camera_pos, camera_size);
 
@@ -35,10 +35,10 @@ fn main() {
             Event::Key(KeyEvent { code: KeyCode::Esc, ..  }) => break,
             Event::Key(KeyEvent { code: kc, .. }) => {
                 match kc {
-                    KeyCode::Left => { player.1.x -= 1.0; }
-                    KeyCode::Right => { player.1.x += 1.0; }
-                    KeyCode::Up => { player.1.y -= 1.0; }
-                    KeyCode::Down => { player.1.y += 1.0; }
+                    KeyCode::Left => { player.1.x -= 1; }
+                    KeyCode::Right => { player.1.x += 1; }
+                    KeyCode::Up => { player.1.y -= 1; }
+                    KeyCode::Down => { player.1.y += 1; }
                     _ => {}
                 }
             }
